% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{col.dens}
\alias{col.dens}
\title{col.dens}
\usage{
col.dens(dens, xlim, fill = "black", border = NULL)
}
\arguments{
\item{dens}{a density object}

\item{xlim}{numeric vector of length 2 specifying the interval}

\item{fill}{color, the color that will fill in the area under the curve, defaults to black}

\item{border}{color, the outline of the area receives this color, defaults to fill}
}
\description{
Original by Arthur Lui, adapted by Mickey
Colors a specified area under a density within an interval by making
a call to the polygon function
}
\examples{
x = rnorm(100)
dens = density(x)
plot(dens)
col.dens(dens, c(0, 1), fill = "blue", border = NA)
}

