% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_fade.R
\name{col_fade}
\alias{col_fade}
\title{col_fade}
\usage{
col_fade(color, alpha)
}
\arguments{
\item{color}{either a string of a valid R color name or a vector of
length 3 of RGB values in [0, 1]^3}

\item{alpha}{the degree of transparency, 0 is invisible and 1 is no
transparency}
}
\description{
Make a color transparent, simple call to rgb()
}
\examples{
col_fade("lightgreen", 0.5)
col_fade(c(0.564, 0.933, 0.564), 0.5)
}
\seealso{
col_dens, col_mult, col_gray
}
