% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_gray.R
\name{col_gray}
\alias{col_gray}
\title{col_gray}
\usage{
col_gray(rgb, method = "yprime")
}
\arguments{
\item{rgb}{numeric vector in [0, 1]^3}

\item{method}{either "yprime", "average", or "lightness". "yprime" attempts to approximate human vision, "average" simply takes the average of rgb and assigns this to all RGB values, "lightness" takes the average of the range of rgb.}
}
\description{
Convert a given RGB color to gray-scale
}
\examples{
col_gray(c(0.8, 0.1, 0.8), "yprime")
}
\seealso{
col_dens, col_mult
}

