% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_mult.R
\name{col_mult}
\alias{col_mult}
\title{col_mult}
\usage{
col_mult(col1, col2)
}
\arguments{
\item{col1}{a color, either in name or integer/hexidecimal}

\item{col2}{a color, either in name or integer/hexidecimal}
}
\description{
Multiplies two colors together.
If A and B are vectors in [0,1]^3 (i.e. {R, G, B}), then to multiply the
colors A and B, do elementwise multiplication.
}
\examples{
col_mult("red", "yellow")
# returns #FF0000, i.e. red

col_mult("dodgerblue", "white")
# returns #1E90FF, i.e. dodgerblue ("white" is identity)

col_mult("dodgerblue", "gray50")
# returns #0F487F, i.e. a darkened version of dodgerblue
}
\seealso{
col_dens, col_gray, col_fade
}
