% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpd_mult.R
\name{hpd_mult}
\alias{hpd_mult}
\title{hpd_mult}
\usage{
hpd_mult(x, dens, prob = 0.95, klength = 5000)
}
\arguments{
\item{x}{vector of samples from some distribution}

\item{dens}{a density object based on x, defaults to density(x)}

\item{prob}{numeric in (0, 1) the probability that the interval should be}

\item{klength}{numeric, determines how accurate the resulting hpd should be,
larger is more precise}
}
\description{
Compute the highest density posterior region from a (possibly multi-modal)
sample of points. If the distribution of the points is known to be unimodal,
then hpd_uni() should be used instead as it will be faster and more accurate
}
\examples{
set.seed(1)
x = c(rnorm(100), rnorm(100, 5)
hpd = hpd_mult(x)
}
\seealso{
hpd_uni
}

