% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpd_uni.R
\name{hpd_uni}
\alias{hpd_uni}
\title{hpd_uni}
\usage{
hpd_uni(x, prob = 0.95, precision = 1000)
}
\arguments{
\item{x}{vector of samples from some distribution}

\item{prob}{numeric in (0, 1) the probability that the interval should be}

\item{precision}{numeric, determines how accurate the resulting hpd should be,
larger is more precise}
}
\description{
Compute the highest density posterior interval from a (unimodal) sample of points.
Unless (slightly) more accurate results for unimodal distributions are deired,
the hpd_mult function should be used instead as it runs faster and can handle
multiple modes.
}
\examples{
x = rnorm(100)
hpd.x = hpd_uni(x)
hpd.x
}
\seealso{
hpd_mult
}

